+++ documentation/examples/todoist/column_format.ts
@@ -42,8 +42,8 @@ const TaskSchema = coda.makeObjectSchema({
       required: true,
     },
   },
-  primary: "name",
-  id: "taskId",
+  displayProperty: "name",
+  idProperty: "taskId",
   identity: {
     name: "Task",
   },