+++ CHANGELOG.md
@@ -1,12 +1,16 @@
 # Changelog
+
 This changelog keeps track of all changes to the packs-sdk. We follow conventions from [keepachangelog](https://keepachangelog.com/en/1.0.0/).
 
 ### Unreleased
+
 - **Breaking Change** Added a validation rule that prevents the usage of varargsParameters for sync table getters
-which are not currently supported in the UI.
+  which are not currently supported in the UI.
+- **Future Breaking Change** The properties `id`, `primary`, and `featured` of object schemas will be renamed to `idProperty`, `displayProperty`, and `featuredProperties`, respectively, to better clarify that their values refer to property names within the schema. During a migration period, you may use either name; the original names are now marked as deprecated and will be removed at a future date, no earlier than May 1, 2022.
 - Added "force" option to `LinkSchema` for `LinkDisplayType.Embed`.
 
 ## [0.9.0] - 2022-03-17
+
 - **Breaking Change** ValueHintType.Url will now create a column of type "Link" instead of "Text".
 - Added ValueHintType.Email for new column type "Email".